refactor: move auth store to stores directory and update import paths
